arm64: dts: allwinner: a64: Add interrupt controller node
In order to give firmware full access to r_intc, Linux needs to avoid
using it in the device tree. Most IRQ lines attached to r_intc are also
attached to the GIC. Unfortunately, the NMI is not. For it to be used in
Linux, we have to forward it in firmware over the mailbox.

Add a node representing this virtual interrupt controller.

Signed-off-by: Samuel Holland <samuel@sholland.org>